requirement.txt
absl-py==2.1.0
asgiref==3.8.1
boto3==1.35.74
botocore==1.35.74
certifi==2024.8.30
charset-normalizer==3.4.0
chex==0.1.87
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
etils==1.11.0
filelock==3.16.1
flax==0.10.2
fsspec==2024.10.0
huggingface-hub==0.26.3
humanize==4.11.0
idna==3.10
importlib_resources==6.4.5
jax==0.4.35
jaxlib==0.4.35
jmespath==1.0.1
markdown-it-py==3.0.0
mdurl==0.1.2
ml_dtypes==0.5.0
msgpack==1.1.0
nest-asyncio==1.6.0
numpy==2.1.3
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.10.1
packaging==24.2
pillow==11.0.0
protobuf==5.29.0
psycopg2==2.9.10
Pygments==2.18.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
safetensors==0.4.5
scipy==1.14.1
sendgrid==6.11.0
simplejson==3.19.3
six==1.16.0
sqlparse==0.5.2
starkbank-ecdsa==2.2.0
tensorstore==0.1.69
tokenizers==0.20.3
toolz==1.0.0
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.21.0
torch==2.5.1